+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 3.16.0
-  - Added request header Elastic-Api-Version
+  - Added request header `Elastic-Api-Version` for serverless [#174](https://github.com/logstash-plugins/logstash-filter-elasticsearch/pull/174)
 
 ## 3.15.2
   - Added checking for `query` and `query_template`. [#171](https://github.com/logstash-plugins/logstash-filter-elasticsearch/pull/171)